requirements.txt
notion-client>=2.0.0
PyYAML>=6.0
